+++ src/libsystemd/sd-login/sd-login.c
@@ -945,11 +945,15 @@ _public_ int sd_machine_get_ifindices(const char *machine, int **ifindices) {
 }
 
 static int MONITOR_TO_FD(sd_login_monitor *m) {
-        return (int) (unsigned long) m - 1;
+        /* Cheri note: see FD_TO_MONITOR */
+        return (int) (uintptr_t) m - 1;
 }
 
 static sd_login_monitor* FD_TO_MONITOR(int fd) {
-        return (sd_login_monitor*) (unsigned long) (fd + 1);
+        /* Cheri note: The pointer returned here is not a real pointer and does
+         * not seem to be used directly, so the cast is ok but results in a
+         * cheri compiler warning. To silence this warning we cast to uintptr_t first. */
+        return (sd_login_monitor*) (uintptr_t) (fd + 1);
 }
 
 _public_ int sd_login_monitor_new(const char *category, sd_login_monitor **m) {
